// src/cds-generated/entities/Account.ts
export const accountMetadata = {
  typeName: "mscrm.account",
  logicalName: "account",
  collectionName: "accounts",
  primaryIdAttribute: "accountid",
  attributeTypes: {
    // Numeric Types
    opendeals: "Integer",
    openrevenue_state: "Integer",
    aging90: "Money",
    address1_longitude: "Double",
    entityimage_timestamp: "BigInt",
    aging90_base: "Money",
    aging60: "Money",
    aging30: "Money",
    aging60_base: "Money",
    aging30_base: "Money",
    msdyn_travelcharge_base: "Money",
    onholdtime: "Integer",
    creditlimit: "Money",
    revenue: "Money",
    exchangerate: "Decimal",
    dev1_calculated: "Integer",
    marketcap: "Money",
    msdyn_travelcharge: "Money",
    opendeals_state: "Integer",
    address2_utcoffset: "Integer",
    utcconversiontimezonecode: "Integer",
    dev1_rollup: "Integer",
    openrevenue: "Money",
    revenue_base: "Money",
    timezoneruleversionnumber: "Integer",
    openrevenue_base: "Money",
    address2_longitude: "Double",
    creditlimit_base: "Money",
    sharesoutstanding: "Integer",
    marketcap_base: "Money",
    address1_utcoffset: "Integer",
    importsequencenumber: "Integer",
    versionnumber: "BigInt",
    address1_latitude: "Double",
    teamsfollowed: "Integer",
    dev1_rollup_state: "Integer",
    numberofemployees: "Integer",
    address2_latitude: "Double",
    // Optionsets
    industrycode: "Optionset",
    dev1_accountlevel: "Optionset",
    msdyn_travelchargetype: "Optionset",
    businesstypecode: "Optionset",
    address1_shippingmethodcode: "Optionset",
    shippingmethodcode: "Optionset",
    accountratingcode: "Optionset",
    preferredappointmenttimecode: "Optionset",
    accountclassificationcode: "Optionset",
    preferredcontactmethodcode: "Optionset",
    address2_freighttermscode: "Optionset",
    statecode: "Optionset",
    address1_addresstypecode: "Optionset",
    paymenttermscode: "Optionset",
    address2_addresstypecode: "Optionset",
    accountcategorycode: "Optionset",
    statuscode: "Optionset",
    customertypecode: "Optionset",
    customersizecode: "Optionset",
    sf365_accounttype: "MultiSelect",
    ownershipcode: "Optionset",
    dev1_multiselect: "MultiSelect",
    preferredappointmentdaycode: "Optionset",
    address2_shippingmethodcode: "Optionset",
    dev1_securedoptionset: "Optionset",
    address1_freighttermscode: "Optionset",
    territorycode: "Optionset",
    // Date Formats
    overriddencreatedon: "DateOnly:UserLocal",
    dev1_datetimeuserlocal: "DateAndTime:UserLocal",
    dev1_dateonlydateonly: "DateOnly:DateOnly",
    createdon: "DateAndTime:UserLocal",
    dev1_datetimetimezoneindependent: "DateAndTime:TimeZoneIndependent",
    lastonholdtime: "DateAndTime:UserLocal",
    openrevenue_date: "DateAndTime:UserLocal",
    dev1_datetimeuserlocalchange: "DateAndTime:UserLocal",
    dev1_dateonlyuserlocal: "DateOnly:UserLocal",
    lastusedincampaign: "DateOnly:UserLocal",
    modifiedon: "DateAndTime:UserLocal",
    opendeals_date: "DateAndTime:UserLocal",
    dev1_rollup_date: "DateAndTime:UserLocal",
    dev1_dateonlytimezoneindependent: "DateOnly:TimeZoneIndependent",
  },
  navigation: {
    transactioncurrencyid: ["mscrm.transactioncurrency"],
    territoryid: ["mscrm.territory"],
    stageid_processstage: ["mscrm.processstage"],
    sla_account_sla: ["mscrm.sla"],
    sf365_SupplierId: ["mscrm.sf365_supplier"],
    sf365_LegalEntityId: ["mscrm.sf365_legalentity"],
    primarycontactid: ["mscrm.contact"],
    preferredsystemuserid: ["mscrm.systemuser"],
    preferredserviceid: ["mscrm.service"],
    preferredequipmentid: ["mscrm.equipment"],
    parentaccountid: ["mscrm.account"],
    owninguser: ["mscrm.systemuser"],
    owningteam: ["mscrm.team"],
    owningbusinessunit: ["mscrm.businessunit"],
    ownerid: ["mscrm.principal"],
    originatingleadid: ["mscrm.lead"],
    msdyn_serviceterritory: ["mscrm.territory"],
    msdyn_salestaxcode: ["mscrm.msdyn_taxcode"],
    msdyn_accountkpiid: ["mscrm.msdyn_accountkpiitem"],
    msdyn_PreferredResource: ["mscrm.bookableresource"],
    msa_managingpartnerid: ["mscrm.account"],
    modifiedonbehalfby: ["mscrm.systemuser"],
    modifiedby: ["mscrm.systemuser"],
    masterid: ["mscrm.account"],
    defaultpricelevelid: ["mscrm.pricelevel"],
    createdonbehalfby: ["mscrm.systemuser"],
    createdby: ["mscrm.systemuser"],
    slainvokedid: ["sla"],
    msdyn_billingaccount: ["account"],
  },
};