// src/cds-generated/entities/Opportunity.ts
export const opportunityMetadata = {
  typeName: "mscrm.opportunity",
  logicalName: "opportunity",
  collectionName: "opportunities",
  primaryIdAttribute: "opportunityid",
  attributeTypes: {
    // Numeric Types
    exchangerate: "Decimal",
    discountamount_base: "Money",
    totalamountlessfreight: "Money",
    totallineitemamount_base: "Money",
    versionnumber: "BigInt",
    discountamount: "Money",
    totallineitemamount: "Money",
    discountpercentage: "Decimal",
    utcconversiontimezonecode: "Integer",
    onholdtime: "Integer",
    actualvalue: "Money",
    teamsfollowed: "Integer",
    freightamount_base: "Money",
    budgetamount: "Money",
    totaltax: "Money",
    totallineitemdiscountamount: "Money",
    importsequencenumber: "Integer",
    totaldiscountamount: "Money",
    totaltax_base: "Money",
    totallineitemdiscountamount_base: "Money",
    totalamountlessfreight_base: "Money",
    estimatedvalue_base: "Money",
    actualvalue_base: "Money",
    estimatedvalue: "Money",
    totalamount: "Money",
    totalamount_base: "Money",
    budgetamount_base: "Money",
    closeprobability: "Integer",
    timezoneruleversionnumber: "Integer",
    freightamount: "Money",
    totaldiscountamount_base: "Money",
    // Optionsets
    timeline: "Optionset",
    msdyn_ordertype: "Optionset",
    prioritycode: "Optionset",
    opportunityratingcode: "Optionset",
    purchasetimeframe: "Optionset",
    statuscode: "Optionset",
    initialcommunication: "Optionset",
    skippricecalculation: "Optionset",
    salesstagecode: "Optionset",
    purchaseprocess: "Optionset",
    need: "Optionset",
    statecode: "Optionset",
    salesstage: "Optionset",
    budgetstatus: "Optionset",
    pricingerrorcode: "Optionset",
    msdyn_forecastcategory: "Optionset",
    // Date Formats
    actualclosedate: "DateOnly:DateOnly",
    overriddencreatedon: "DateOnly:UserLocal",
    schedulefollowup_qualify: "DateOnly:UserLocal",
    finaldecisiondate: "DateOnly:DateOnly",
    createdon: "DateAndTime:UserLocal",
    schedulefollowup_prospect: "DateOnly:UserLocal",
    estimatedclosedate: "DateOnly:DateOnly",
    modifiedon: "DateAndTime:UserLocal",
    lastonholdtime: "DateAndTime:UserLocal",
    scheduleproposalmeeting: "DateOnly:UserLocal",
  },
  navigation: {
    transactioncurrencyid: ["mscrm.transactioncurrency"],
    stageid_processstage: ["mscrm.processstage"],
    sla_opportunity_sla: ["mscrm.sla"],
    pricelevelid: ["mscrm.pricelevel"],
    parentcontactid: ["mscrm.contact"],
    parentaccountid: ["mscrm.account"],
    owninguser: ["mscrm.systemuser"],
    owningteam: ["mscrm.team"],
    owningbusinessunit: ["mscrm.businessunit"],
    ownerid: ["mscrm.principal"],
    originatingleadid: ["mscrm.lead"],
    msdyn_workordertype: ["mscrm.msdyn_workordertype"],
    msdyn_opportunitykpiid: ["mscrm.msdyn_opportunitykpiitem"],
    msdyn_ContractOrganizationalUnitId: ["mscrm.msdyn_organizationalunit"],
    msdyn_AccountManagerId: ["mscrm.systemuser"],
    msa_partneroppid: ["mscrm.contact"],
    msa_partnerid: ["mscrm.account"],
    modifiedonbehalfby: ["mscrm.systemuser"],
    modifiedby: ["mscrm.systemuser"],
    createdonbehalfby: ["mscrm.systemuser"],
    createdby: ["mscrm.systemuser"],
    campaignid: ["mscrm.campaign"],
    customerid: ["account","contact"],
    slainvokedid: ["sla"],
  },
};