// src/cds-generated/entities/OpportunityClose.ts
export const opportunitycloseMetadata = {
  typeName: "mscrm.opportunityclose",
  logicalName: "opportunityclose",
  collectionName: "opportunitycloses",
  primaryIdAttribute: "activityid",
  attributeTypes: {
    // Numeric Types
    actualrevenue_base: "Money",
    utcconversiontimezonecode: "Integer",
    scheduleddurationminutes: "Integer",
    onholdtime: "Integer",
    timezoneruleversionnumber: "Integer",
    exchangerate: "Decimal",
    importsequencenumber: "Integer",
    actualrevenue: "Money",
    actualdurationminutes: "Integer",
    versionnumber: "BigInt",
    // Optionsets
    instancetypecode: "Optionset",
    opportunitystatuscode: "Optionset",
    deliveryprioritycode: "Optionset",
    statecode: "Optionset",
    community: "Optionset",
    prioritycode: "Optionset",
    statuscode: "Optionset",
    opportunitystatecode: "Optionset",
    // Date Formats
    modifiedon: "DateAndTime:UserLocal",
    postponeactivityprocessinguntil: "DateAndTime:UserLocal",
    deliverylastattemptedon: "DateAndTime:UserLocal",
    scheduledstart: "DateOnly:UserLocal",
    sortdate: "DateAndTime:UserLocal",
    lastonholdtime: "DateAndTime:UserLocal",
    actualend: "DateOnly:UserLocal",
    overriddencreatedon: "DateOnly:UserLocal",
    createdon: "DateAndTime:UserLocal",
    actualstart: "DateOnly:UserLocal",
    senton: "DateAndTime:UserLocal",
    scheduledend: "DateOnly:UserLocal",
  },
  navigation: {
    sla_activitypointer_sla_opportunityclose: ["mscrm.sla"],
    ownerid_opportunityclose: ["mscrm.principal"],
    opportunityid: ["mscrm.opportunity"],
    competitorid: ["mscrm.competitor"],
    activityid_activitypointer: ["mscrm.activitypointer"],
    regardingobjectid: ["account","bookableresourcebooking","bookableresourcebookingheader","bulkoperation","campaign","campaignactivity","entitlement","entitlementtemplate","incident","interactionforemail","knowledgearticle","knowledgebaserecord","lead","opportunity","site"],
    transactioncurrencyid: ["transactioncurrency"],
    slainvokedid: ["sla"],
    createdby: ["systemuser"],
    modifiedby: ["systemuser"],
    serviceid: ["service"],
    owninguser: ["systemuser"],
    modifiedonbehalfby: ["systemuser"],
    owningteam: ["team"],
    owningbusinessunit: ["businessunit"],
    sendermailboxid: ["mailbox"],
    createdonbehalfby: ["systemuser"],
  },
};